requirements.txt
python-telegram-bot[webhooks]
python-dotenv
praw
pymongo
pydantic